Add amusement-park - Iteration 1
Practice JS object handling.

Introduced two new methods.

Optional chaining `?` - tries to access a nested key if the parent is
not `null` or `undefined` and if it is then returns `undefined`.

Nullish Coalescing `??` - It applies a default value to a variable if
and only if the variable is `null` or `undefined`.