Add coordinate-transformation - Iteration 3
Cleanup and compress the functions.

It's incredible how tight the arrow function definition can get.

`composeTransform` line 48: `...` is the `spread` syntax operator. It
allows an iterable to be expanded in places where zero or more arguments
or elements are expected, or an object expression to be expanded in
places where zero or more key-value pairs are expected.

Here this operator takes the array that `f(x,y)` returns and uses that
as input coordinates for the second transform.