Add coordinate-transformation - Iteration 5
Simplified memoization. The input and the result coordinates are no
longer stored in an `Object`.

Instead we have a simple variable for each of them. This removes the
need to reset the object to keep it from storing more than the latest
transform.

I ended up compressing the function beyond need and made it a bit messy
to read.

diff --git a/coordinate-transformation.js b/coordinate-transformation.js
index 3720ac1..55dc46e 100644
--- a/coordinate-transformation.js
+++ b/coordinate-transformation.js
@@ -58,15 +58,12 @@ import { cp } from "fs"
  *  if the arguments are the same on subsequent calls, or compute a new result if they are different.
  */
  export function memoizeTransform(f) {
-  let cache = {}
+  let cachedCoords, cachedResult
   return (x, y) => {
-    let key = `${[x, y]}`
-    if (key in cache) {
-      return cache[key]
+    if (cachedCoords === `${[x, y]}`) {
+      return cachedResult
     }
-    let result = f(x, y)
-    cache = {}
-    cache[key] = result
-    return result
+    cachedCoords = `${[x, y]}`
+    return cachedResult = f(x, y)
   }
 }