Add coordinate-transformation - Iteration 7
According to my mentor this was a good example where it's better to use
a few more lines than less. The code did get a bit clogged.

The memoized input and result are now defined on their own lines with a
default value assigned to them. These default values are something that
these variables should never hold, a string.

Having each variable declaration on their own line makes them fast to
read and obvious. It's easy not to really see a multiple variable
declaration.

Renamed most of the variables to better describe them.