Add fruit-picker - Iteration 1
Using function declarations and most typed-out forms of expressions to
self-comment.

`isServiceOnline` A function declaration for the callback that takes in
the string that tells the status of the service. This is then fed into
`checkStatus` function that's `imported` from `./grocer` as a callback.

`pickFruit` takes in the args `variety` and `quantity` as well as a
callback. The `checkInventory` `imported` from `./grocer` takes in a
`query _Object_` that need to be formed from the arguments. The object
creation can and will be shortened in following iterations.

`purchaseInventoryIfAvailable` has args `err` as a string and
`isAvailable` as a boolean. If there's anything more than null in `err`
we have an error thrown at us. If the `err` is `null` and the item is
available 'PURCHASE' is returned. If isn't we get 'NOOP'. Can and will
be shortened.

´pickAndPurchaseFruit` takes in args for the desired fruit ´variety´ and
´quantity´. They're then thrown into ´pickFruit´ together with
´purchaseInventoryIfAvailable´ as a callback.