Add elyses-transformative-enchantments - Iteration 1
Learning exercise of array transformations.

`threeOfEachThree` was a bitch and still smells. First step is to dig
out the indexes of the threes in the deck if any. That's done with `map`
here. Inserting the extra threes into the deck pushes the indexes
forward too so it's better to start from the end with `splice` which
allows also adding elements to the array. Finally the deck is returned.

`sandwichTrick` is also a bit smelly. The first and last cards are taken
out of the deck and stored. Middle of the deck is calculated and then
the removed cards are inserted into the middle of the deck in reverse
order.

`perfectlyOrdered` needed more than just `sort`, because just `sort`
resulted in [1, 10, 2, 3, ..]. Good to know this.