Add elyses-transformative-enchantments - Iteration 2
`threeOfEachThree` This can also be done with `reduce`, which I didn't
even consider at first. Into the accumulator array named `filledDeck` we
push the cards unless it's a three we push thrice that three.

`perfectlyOrdered` My first iteration was a massive Spacehog. This is
rediculously cleaner and shorter. Someday I'll remember these when
sorting.