Add windowing-system - Iteration 1
I made an effort and created the comments in the file too.

The limits for moving the window took a bit of head scratching. My new
chalkboard helped with that.

I'm looking forward to my mentor's comments on this one. I liked this
exercise. I felt like really building something.