Add factory-sensors - Iteration 1
Error handling exercise.

`OverheatingError` has that `temperature` attribute that is accessed
like any object attribute. Theres no need to have and carry that in the
monitoring function. It's just `error.temperature`.

For `actions` I first had full function declarations outside of the
object, but the found that it can be simplified significantly with empty
arrow functions.

I also thought that I needed to have that `UnknownError` extension in
the file, but it's coming in from the test file.