Add pangram - Iteration 2
Here we have the ALPHABET formed with `split('')` instead of
`Array.from`. I'm not sure which is better for performance. I kinda do
like the `Array.from` when it comes to readability.

For performance on very long input sentences I've introduced the const
`lowerCaseSentence` so that the sentence isn't run through `toLowerCase`
for every char check in the sencence. Although the check is against the
`ALPHABET` which has a fixed number of chars.