Add ozans-playlist - Iteration 3
Using `removeDuplicates` which gives us array of the Set to work with.

`hasTrack` now checks the array with `includes()`. Looks like there's
some unnecessary conversions happening when there's the `Set` method
`has()` available.

`addTrack` now spreads the playlist together with the track to add and
then throws that into `removeDuplicates` before spreading the result.

`deleteTrack` now filters the playlist for the track to be deleted
before creating the Set and spreading it out.

`listArtists` now creates a Set using `removeDuplicates`