Add matrix - Iteration 2
`rows` prettied the dot notation by breaking lines before the dot for
function calls.

`columns` Because `this.rows` is a function and it gets called on every
iteration cycle, it's extra work. Created a const to hold the rows and
working on that instead. The `for` loop got a major overhaul too.
Instead of `push()`ing into an intermediate variable with `forEach()` we
use `map` to get us and array for each row's element in turn and only
that gets pushed into the columns array.