Add collatz-conjecture - Iteration 4
Jumped to the recursive bandwagon.

Now we throw the number back into `steps` after doing the conjecture on
the number. Count is declared as a parameter with default value and
increased on every cycle. Once n reaches 1 it returns the count.

It's pretty sweet. I didn't know I could cycle something through one
function like this.