Add matrix - Iteration 4
`rows()` I got to take out one `map` by moving the mapping to numbers to
follow `split`. There we're already dealing with individual elements of
each row.

`columns()` It took a while to figure that at first we only need the
indexes from a row in the matrix to create our columns. Then we map
through each row with the index to have our columns returned.