Add raindrops - Iteration 1
A constant Object to hold the SOUNDS.

First iterate over the keys in SOUNDS to check if any of them are a
factor of our input n.

If there are any `map` them to their values and `join` them together. If
there isn't any we return the n converted `toString`