Add phone-number - Iteration 2
Instead of that `if` hell in iteration one, this is pretty clean!

In `VALIDATIONS` we have and array of arrays holding pairs of function
followed by an error message.

In function `clean` we first process the input by replacing
non-numericals with nothing using `replace`. Then it's normalized by
adding the country code '1' to the beginning if needed.

Iterating through `VALIDATIONS` checks if an error condition is met and
if it is it throws an error and if not the number is returned without
the NA country code '1'.