Add dnd-character - Iteration 1
This static came as new. Static attributes for an instance of the class
are calculated only once upon creation.

Static properties and methods are called without instantiating their
class and cannot be called through a class instance. Static methods are
often used to create utility functions for an application, where as
static properties are useful for caches, fixed-configuration, or any
other data you don't need to be replicated across instances.