Add linked-list - Iteration 2
Remove this.length and cumulative upkeep of it. Instead the count is now
actually counted when it's called.

Refactor Node to be a standalone class instead of a prototype of
LinkedList, because there's no need for it to inherit anything.

Refactor pop to not break when popping to empty witout having
this.length.